fastapi_app/requirements.txt
# FastAPI and core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP and networking
httpx>=0.25.0
aiofiles>=23.0.0

# AI and Machine Learning (optional)
# openai>=1.0.0
# langchain>=0.0.300
# tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document Processing (optional)
# PyPDF2>=3.0.0
# python-docx>=1.0.0
# Pillow>=10.0.0

# Email
# fastapi-mail>=1.4.0
jinja2>=3.1.0

# File handling and utilities
# python-magic>=0.4.0
python-dateutil>=2.8.0
python-slugify>=8.0.0

# Background tasks (optional - for future use)
# celery>=5.3.0
# redis>=5.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0

# Production
gunicorn>=21.0.0

# Data validation
email-validator>=2.0.0

# Utilities
pytz>=2023.3